+++ src/Defaults.ts
@@ -1,8 +1,10 @@
-/**
- * @format
- **/
+/** @format */
+
+import { Preset } from './Types';
 
 export const DEFAULT_LINE_WIDTH = 2;
 export const DEFAULT_LINE_COLOR = '#000';
 export const DEFAULT_CIRCLE_COLOR = '#000';
 export const DEFAULT_DOT_COLOR = '#FFF';
+export const DEFAULT_END_WITH_CIRCLE = false;
+export const DEFAULT_PRESET: Preset = 'SingleColumnLeft';